Restructure packaging for issuer-list
COAND-497
