+++ entercash/src/main/java/com/adyen/checkout/entercash/EntercashComponent.kt
@@ -15,7 +15,7 @@ import com.adyen.checkout.components.model.payments.request.EntercashPaymentMeth
 import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.entercash.EntercashComponent.Companion.PROVIDER
 import com.adyen.checkout.issuerlist.IssuerListComponent
-import com.adyen.checkout.issuerlist.IssuerListDelegate
+import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
 
 /**
  * Component should not be instantiated directly. Instead use the [PROVIDER] object.
