+++ issuer-list/src/main/java/com/adyen/checkout/issuerlist/internal/ui/IssuerListDelegate.kt
@@ -6,16 +6,22 @@
  * Created by oscars on 11/7/2022.
  */
 
-package com.adyen.checkout.issuerlist
+package com.adyen.checkout.issuerlist.internal.ui
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
 import com.adyen.checkout.components.model.payments.request.IssuerListPaymentMethod
 import com.adyen.checkout.components.ui.ButtonDelegate
 import com.adyen.checkout.components.ui.UIStateDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
+import com.adyen.checkout.issuerlist.internal.ui.model.IssuerListComponentParams
+import com.adyen.checkout.issuerlist.internal.ui.model.IssuerListInputData
+import com.adyen.checkout.issuerlist.internal.ui.model.IssuerListOutputData
+import com.adyen.checkout.issuerlist.internal.ui.model.IssuerModel
 import kotlinx.coroutines.flow.Flow
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 interface IssuerListDelegate<IssuerListPaymentMethodT : IssuerListPaymentMethod> :
     PaymentComponentDelegate<PaymentComponentState<IssuerListPaymentMethodT>>,
     ViewProvidingDelegate,
