+++ issuer-list/src/main/java/com/adyen/checkout/issuerlist/internal/ui/model/IssuerListOutputData.kt
@@ -5,11 +5,13 @@
  *
  * Created by caiof on 25/4/2019.
  */
-package com.adyen.checkout.issuerlist
+package com.adyen.checkout.issuerlist.internal.ui.model
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.OutputData
 
-class IssuerListOutputData(
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+data class IssuerListOutputData(
     val selectedIssuer: IssuerModel?,
 ) : OutputData {
